# src/lib/supabase-inspect/openapi.yaml
openapi: 3.1.0
info:
  title: Supabase Inspect Action
  version: 1.0.0
  description: >
    Action read-only do LP Factory 10 para inspeções estruturadas no Supabase.

servers:
  - url: https://REPLACE_WITH_APP_URL

security:
  - AgentKeyAuth: []

paths:
  /api/agent-tools/supabase-inspect:
    post:
      operationId: run_supabase_inspect_readonly
      summary: Executa inspeção read-only no Supabase
      description: >
        Recebe um pedido estruturado de inspeção, consulta o Supabase em modo
        somente leitura e devolve output organizado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - inspectionType
                - target
              properties:
                inspectionType:
                  type: string
                  enum:
                    - tableOverview
                    - schemaColumns
                    - rlsPolicies
                    - indexes
                    - views
                    - functions
                    - triggers
                    - dataSample
                target:
                  type: object
                  required:
                    - schema
                    - kind
                  properties:
                    schema:
                      type: string
                    name:
                      type: string
                    kind:
                      type: string
                      enum:
                        - table
                        - view
                        - function
                        - schema
                questions:
                  type: array
                  items:
                    type: string
                rowLimit:
                  type: integer
                  minimum: 1
                  maximum: 50
                responseMode:
                  type: string
                  enum:
                    - structuredReport
      responses:
        "200":
          description: Inspeção executada com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectReadonlyOutput"
        "400":
          description: Input inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectReadonlyOutput"
        "422":
          description: Escopo bloqueado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectReadonlyOutput"
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectReadonlyOutput"

components:
  securitySchemes:
    AgentKeyAuth:
      type: apiKey
      in: header
      name: x-agent-key

  schemas:
    InspectFinding:
      type: object
      required:
        - title
        - severity
        - detail
      properties:
        title:
          type: string
        severity:
          type: string
          enum:
            - info
            - attention
            - blocked
        detail:
          type: string
        evidence:
          type: array
          items:
            type: string
        payload:
          type: array
          items:
            type: object
            additionalProperties: true

    InspectReadonlyOutput:
      type: object
      required:
        - status
        - requestSummary
        - normalizedScope
        - findings
        - inspectedObjects
        - warnings
        - limitations
        - nextSteps
      properties:
        status:
          type: string
          enum:
            - ok
            - needs_clarification
            - blocked
            - error
        requestSummary:
          type: string
        normalizedScope:
          oneOf:
            - type: "null"
            - type: object
              additionalProperties: true
        findings:
          type: array
          items:
            $ref: "#/components/schemas/InspectFinding"
        inspectedObjects:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        limitations:
          type: array
          items:
            type: string
        nextSteps:
          type: array
          items:
            type: string